+++ Changelog.md
@@ -8,6 +8,7 @@
   on poisoning
 * Stay in bank, monastery and school until explicitly leaving (with ESC)
 * Configure game defaults with CTRL-P
+* Show a popup window during actions that take more than ten turns
 * Show item type description in auto-pickup configuration dialogue
 * Permanently show spell description below spell selection window
 * Describe spell type for most spells in spell description